Clean up project structure and consolidate files
