+++ package.json
@@ -5,11 +5,10 @@
   "main": "backend/server.js",
   "type": "module",
   "scripts": {
-    "dev": "concurrently \"npm run dev:frontend\" \"npm run dev:backend\"",
-    "dev:frontend": "vite --config frontend/vite.config.js",
+    "dev": "concurrently \"vite\" \"npm run dev:backend\"",
     "dev:backend": "node backend/server.js",
-    "build": "vite build --config frontend/vite.config.js",
-    "preview": "vite preview --config frontend/vite.config.js",
+    "build": "vite build",
+    "preview": "vite preview",
     "start": "node backend/server.js"
   },
   "keywords": ["emergency", "call-center", "elevenlabs", "voice-ai"],
@@ -17,6 +16,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@elevenlabs/client": "0.1.7",
+    "@supabase/supabase-js": "^2.39.0",
     "express": "^4.18.2",
     "cors": "^2.8.5"
   },
